+++ AutoPkgr/LGNotificationManager.h
@@ -1,9 +1,6 @@
+//  LGNotificationManager.h
 //
-//  LGNotificationsViewController.h
-//  AutoPkgr
-//
-//  Created by Eldon on 5/20/15.
-//  Copyright 2015 Eldon Ahrold.
+//  Copyright 2015 Eldon Ahrold
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -15,11 +12,16 @@
 //  distributed under the License is distributed on an "AS IS" BASIS,
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
-//  limitations under the License.//
+//  limitations under the License.
+//
 
-#import <Cocoa/Cocoa.h>
-#import "LGTabViewControllerBase.h"
+#import <Foundation/Foundation.h>
+@interface LGNotificationManager : NSObject
 
-@interface LGNotificationsViewController : LGTabViewControllerBase
+@property (copy, nonatomic, readonly) NSDictionary *reportDictionary;
+- (instancetype)init NS_UNAVAILABLE;
+- (instancetype)initWithReportDictionary:(NSDictionary *)dictionary errors:(NSError *)error;
+
+- (void)sendEnabledNotifications:(void (^)(NSError *error))complete;
 
 @end