+++ AutoPkgr/LGNotificationsViewController.m
@@ -18,12 +18,39 @@
 //  limitations under the License.//
 
 #import "LGNotificationsViewController.h"
-#import "LGPasswords.h"
 #import "LGAutoPkgr.h"
+#import "LGPasswords.h"
 #import "LGTestPort.h"
-#import "LGEmailer.h"
+#import "LGEmailNotification.h"
+#import "LGSlackNotification.h"
 
 @interface LGNotificationsViewController ()
+#pragma mark - Email
+#pragma mark-- Outlets --
+
+@property (weak) IBOutlet NSTextField *smtpServer;
+@property (weak) IBOutlet NSTextField *smtpPort;
+
+@property (weak) IBOutlet NSTextField *smtpUsername;
+@property (weak) IBOutlet NSSecureTextField *smtpPassword;
+
+@property (weak) IBOutlet NSTextField *slackWebhookURLTF;
+@property (weak) IBOutlet NSProgressIndicator *slackProgressIndicator;
+@property (weak) IBOutlet NSButton *slackHelpButton;
+
+// Status icons
+@property (weak) IBOutlet NSImageView *testSmtpServerStatus;
+
+// Spinners
+@property (weak) IBOutlet NSProgressIndicator *sendTestEmailSpinner;
+@property (weak) IBOutlet NSProgressIndicator *testSmtpServerSpinner;
+
+#pragma mark-- IBActions --
+- (IBAction)sendTestEmail:(id)sender;
+- (IBAction)testServerPort:(id)sender;
+
+- (IBAction)getKeychainPassword:(NSTextField *)sender;
+- (IBAction)updateKeychainPassword:(id)sender;
 
 @end
 
@@ -48,6 +75,17 @@
     }];
 
     [self getKeychainPassword:_smtpPassword];
+
+    [LGSlackNotification infoFromKeychain:^(NSString *infoOrPassword, NSError *error) {
+        _slackWebhookURLTF.safe_stringValue = infoOrPassword;
+    }];
+
+
+    [NSTimer timerWithTimeInterval:5
+                            target:[LGPasswords class]
+                          selector:@selector(lockKeychain)
+                          userInfo:nil
+                           repeats:NO];
 }
 
 - (NSString *)tabLabel
@@ -58,16 +96,13 @@
 #pragma mark - Keychain Actions
 - (void)getKeychainPassword:(NSTextField *)sender
 {
-    NSString *account = _smtpUsername.stringValue;
-    if (account.length) {
-        [LGPasswords getPasswordForAccount:account reply:^(NSString *password, NSError *error) {
-            if (error) {
-                NSLog(@"Error getting password for %@ [%ld]: %@", account, error.code, error.localizedDescription);
-            } else {
-                _smtpPassword.safe_stringValue = password;
-            }
-        }];
-    }
+    [LGEmailNotification infoFromKeychain:^(NSString *infoOrPassword, NSError *error) {
+        if (error) {
+            NSLog(@"Error getting password for %@ [%ld]: %@", [LGEmailNotification account], error.code, error.localizedDescription);
+        } else {
+            _smtpPassword.safe_stringValue = infoOrPassword;
+        }
+    }];
 }
 
 - (IBAction)updateKeychainPassword:(id)sender
@@ -80,12 +115,11 @@
     NSString *account = _smtpUsername.safe_stringValue;
     NSString *password = _smtpPassword.safe_stringValue;
 
-    if (account && password) {
-        [LGPasswords savePassword:password forAccount:account reply:^(NSError *error) {
+    if (account) {
+        [LGEmailNotification saveInfoToKeychain:password reply:^(NSError *error) {
             if (reply) {
                 reply(error);
             }
-
             if (error) {
                 if (error.code == errSecAuthFailed || error.code == errSecDuplicateKeychain) {
                     [LGPasswords resetKeychainPrompt:^(NSError *error) {
@@ -144,7 +178,7 @@
     [_sendTestEmailSpinner startAnimation:self]; // animate spinner
 
     // Setup a completion block
-    void (^didComplete)() = ^void(NSError *error) {
+    void (^didComplete)(NSError *) = ^void(NSError *error) {
 //        [_sendTestEmailButton setEnabled:YES]; // enable button
         [_sendTestEmailSpinner setHidden:YES]; // hide spinner
         [_sendTestEmailSpinner stopAnimation:self]; // stop animation
@@ -160,13 +194,8 @@
         }
 
         // Create an instance of the LGEmailer class
-        LGEmailer *emailer = [[LGEmailer alloc] init];
-
-        // Set the reply block
-        emailer.replyBlock = didComplete;
-
-        // Send the test email.
-        [emailer sendTestEmail];
+        LGEmailNotification *emailer = [[LGEmailNotification alloc] init];
+        [emailer sendTest:didComplete];
     }];
 }
 
@@ -192,4 +221,37 @@
     }
 }
 
+- (IBAction)testSlackWebhook:(NSButton *)sender
+{
+    NSString *webHook = _slackWebhookURLTF.stringValue;
+
+    [_slackProgressIndicator startAnimation:self];
+
+    _slackHelpButton.hidden = YES;
+    _slackWebhookURLTF.enabled = NO;
+    sender.enabled = NO;
+
+    void (^stopProgress)(NSError *) = ^(NSError *error){
+        sender.enabled = YES;
+        _slackHelpButton.hidden = NO;
+        _slackWebhookURLTF.enabled = YES;
+        [_slackProgressIndicator stopAnimation:self];
+        [self.progressDelegate stopProgress:error];
+    };
+
+    [LGSlackNotification saveInfoToKeychain:webHook reply:^(NSError *error) {
+        if (error) {
+            stopProgress(error);
+        } else {
+            LGSlackNotification *notification = [[LGSlackNotification alloc] init];
+            [notification sendTest:^(NSError *error) {
+                stopProgress(error);
+            }];
+        }
+    }];
+}
+
+- (IBAction)changeWebHookURL:(NSTextField *)sender {
+    [LGSlackNotification saveInfoToKeychain:sender.stringValue reply:^(NSError *error){}];
+}
 @end